Update ChatAnswerResponse, added get usage meta data and topic, set chat title as a topic

+++ src/main/java/kz/sdu/chat/mainservice/entities/UserTokens.java
@@ -20,7 +20,7 @@ public class UserTokens {
 
     private Long userId;
 
-    private Long spentCost;
+    private Double spentCost;
 
     private LocalDateTime updateDate;
